// src/app/srvapi/models/ProcessorOperationArgument.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProcessorOperationArgument
 */
export interface ProcessorOperationArgument {
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationArgument
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationArgument
     */
    value?: string;
}

export function ProcessorOperationArgumentFromJSON(json: any): ProcessorOperationArgument {
    return ProcessorOperationArgumentFromJSONTyped(json, false);
}

export function ProcessorOperationArgumentFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProcessorOperationArgument {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': !exists(json, 'name') ? undefined : json['name'],
        'value': !exists(json, 'value') ? undefined : json['value'],
    };
}

export function ProcessorOperationArgumentToJSON(value?: ProcessorOperationArgument | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'value': value.value,
    };
}