// src/app/srvapi/models/processor-operation-argument.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ProcessorOperationArgument
 */
export interface ProcessorOperationArgument {
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationArgument
     */
    name?: any;
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationArgument
     */
    value?: any;
}